bouncebox: fullscreen clearing, fixed draw offsets
I switched back to fullscreen clearing, as it was much easier to maintain
than the incremental clearing. I did not do any benchmarks yet so I can't
tell whether it's faster to calculate the lines and so and overdraw it or
to just clear the terminal. It is safer anyway at all.

The way the offets are calculated changed now too. The box border is now
more like a CSS outline, meaning that it's around the box, not belonging
to the box itself. So when the box as an object is at 0,0, it means that
the border of the box will be drawn in the upper left corner of the ter-
minal.

The screen dimensions stored are therefore reduced by the amount of space
needed by the status bar and the box borders. Also, the whole coordinates
system (used on the terminal) is shifted.

The issue with the balls getting drawn in the box border doesn't seem to
be fixed yet (it's explained at the bottom of the script).