bouncebox: collisions revert the movement into the correct direction
Previously, if the boundaries were exceeded, the movement direction was
simply inverted. However, with large delays between simulation steps, it
was possible that the box wasn't moved into the boundaries until the next
step anymore. So, in the next step, the direction would be inverted again
and the box would get stuck where it is. This is now fixed - the direction
is always set to have the correct sign now.

Additionally, the new key r is bound to moving the box near the origin.